Interface and documentation for user input Parser
