+++ web-crawler-frontend/src/main/java/pl/edu/agh/student/oop/webcrawler/frontend/input/Parser.java
@@ -1,72 +1,61 @@
 package pl.edu.agh.student.oop.webcrawler.frontend.input;
 
 import pl.edu.agh.student.oop.webcrawler.core.configuration.Configuration;
-import pl.edu.agh.student.oop.webcrawler.core.configuration.ConfigurationBuilder;
 import pl.edu.agh.student.oop.webcrawler.core.matcher.Matcher;
-import pl.edu.agh.student.oop.webcrawler.core.matcher.MatcherCompiler;
 import pl.edu.agh.student.oop.webcrawler.core.matcher.Matchers;
-import pl.edu.agh.student.oop.webcrawler.core.parser.Word;
 import pl.edu.agh.student.oop.webcrawler.frontend.views.configuration.model.ConditionsListItem;
 
-import java.util.ArrayList;
 import java.util.List;
 
-public class Parser {
+public interface Parser {
 
-    private static final String WILD_CARD = "*";
-    private static final String SPACE_REGEX = "\\s+";
+    /**
+     * Creates new {@link Configuration} object
+     *
+     * @param depth - crawling depth - means how many times crawler should go deeper during searching
+     * @param domains - list of domains crawler will check
+     * @param items - list of {@link ConditionsListItem}
+     * @return new Configuration {@link Configuration}
+     */
+    Configuration createConfiguration(List<ConditionsListItem> items, List<String> domains, String depth);
 
-    public Configuration createConfiguration(String depth, List<String> domains, List<ConditionsListItem> items) {
+    /**
+     * Combines list of {@link Matcher} objects into one {@link Matcher} object
+     *
+     * @param matchers - list of {@link Matcher} objects to combine into one {@link Matcher}
+     * @return new {@link Matcher} object
+     */
+    Matcher combineMatchers(List<Matcher> matchers);
 
-        ConfigurationBuilder builder = Configuration.builder();
-        builder.matcher(createMatcher(items))
-               .domains(domains)
-               .depth(Integer.parseInt(depth));
+    /**
+     * Parses list of {@link ConditionsListItem} into list of {@link Matcher}
+     * using {@link #parseCondition(ConditionsListItem)} method
+     *
+     * @param items list of {@link ConditionsListItem} objects
+     * @return list of {@link Matcher} objects
+     */
+    List<Matcher> parseConditions(List<ConditionsListItem> items);
 
-        return builder.build();
-    }
+    /**
+     * Parse one {@link ConditionsListItem} object into {@link Matcher} object
+     *
+     * @param conditionItem - {@link ConditionsListItem} condition to be parsed into Matcher
+     * @return new {@link Matcher} object
+     */
+    Matcher parseCondition(ConditionsListItem conditionItem);
 
-    private Matcher createMatcher(List<ConditionsListItem> items) {
-
-        List<Matcher> subMatchers = new ArrayList<>();
-        for(ConditionsListItem item : items) {
-            subMatchers.add(parseCondition(item));
-        }
-
-        return Matchers.matchAny(subMatchers.toArray(new Matcher[0]));
-    }
-
-    private Matcher parseCondition(ConditionsListItem conditionItem) {
-        String posSubCondition = conditionItem.getPositiveSubCondition();
-        String negSubCondition = cleanNegation(conditionItem.getNegativeSubCondition());
-
-        Matcher positiveMatcher = parseSubCondition(posSubCondition);
-        Matcher negativeMatcher = Matchers.negate(parseSubCondition(negSubCondition));
-
-        return Matchers.matchEvery(positiveMatcher, negativeMatcher);
-    }
-
-    private Matcher parseSubCondition(String subCondition) {
-        int wildCardCounter = 0;
-        MatcherCompiler matcher = Matcher.compiler().matchAny();
-        for (String string : subCondition.split(SPACE_REGEX)) {
-            if (string.equals(WILD_CARD)) wildCardCounter++;
-            else if (wildCardCounter > 0) {
-                matcher.thenSkip(wildCardCounter);
-                matcher.thenMatch(Word.of(string));
-                wildCardCounter = 0;
-            } else {
-                matcher.thenMatch(Word.of(string));
-            }
-        }
-
-        return matcher.compile();
-
-    }
-
-    private String cleanNegation(String negSubCondition) {
-        return negSubCondition.substring(2, negSubCondition.length()-2);
-    }
-
+    /**
+     * Parses subCondition into {@link Matcher} object.
+     * subCondtion is a simple String like "ala ma * kota".
+     * This string can contain only letters and "*". If you want to get Matcher
+     * that represent negative subCondition you have to use {@link Matchers#negate(Matcher)}
+     * method on result of {@link #parseSubCondition(String)} method
+     *
+     * @param subCondition - String representing positive part of {@link ConditionsListItem}.
+     *                     It can contain only letters and "*"
+     *
+     * @return new {@link Matcher} object
+     */
+    Matcher parseSubCondition(String subCondition);
 
 }
