+++ web-crawler-frontend/src/main/java/pl/edu/agh/student/oop/webcrawler/frontend/views/configuration/presenter/ConfigurationPresenter.java
@@ -7,7 +7,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import pl.edu.agh.student.oop.webcrawler.core.configuration.Configuration;
-import pl.edu.agh.student.oop.webcrawler.frontend.input.Parser;
+import pl.edu.agh.student.oop.webcrawler.frontend.input.InputParser;
 import pl.edu.agh.student.oop.webcrawler.frontend.views.configuration.model.ConditionsListItem;
 
 import java.util.ArrayList;
@@ -99,13 +99,13 @@ public class ConfigurationPresenter {
 
     @FXML
     private void handleSearchAction(ActionEvent event) {
-        Configuration configuration = new Parser().createConfiguration(
-                depthTextField.getText(),
-                new ArrayList<String>(domains),
+        Configuration configuration = new InputParser().createConfiguration(
                 new ArrayList<ConditionsListItem>(listController
                                                   .getConditionsListView()
-                                                  .getItems()));
-        
+                                                  .getItems()),
+                new ArrayList<String>(domains),
+                depthTextField.getText());
+
         this.tabPane.getSelectionModel().select(1);
     }
